# api/_lib/chatgpt-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IMPRIMERIE OGOOUÉ — pont de lecture",
    "version": "1.0.0",
    "description": "Accès en LECTURE SEULE aux données de gestion de l'IMPRIMERIE OGOOUÉ (Moanda, Gabon) : chiffre d'affaires, caisse, commandes, stock, clients, factures, employés, trésorerie. Aucune écriture n'est possible : toute requête autre que GET est refusée par le serveur. Tous les montants sont en francs CFA (XAF). Toutes les dates et heures sont celles de Moanda (fuseau Africa/Libreville, UTC+1 toute l'année) — jamais celles du serveur. Chaque réponse porte un champ `calcule_le` : cite-le quand tu donnes un chiffre, par exemple « au 18/09 à 18 h ». Si une information n'est pas dans une réponse, dis-le plutôt que de l'estimer : ces chiffres servent à décider, pas à illustrer."
  },
  "servers": [
    {
      "url": "https://imprimerie-ogooue-app.vercel.app",
      "description": "Production"
    }
  ],
  "security": [{ "jetonDuPont": [] }],
  "paths": {
    "/api/chatgpt-ca-mois": {
      "get": {
        "operationId": "chiffreAffairesDuMois",
        "summary": "Chiffre d'affaires du mois (facturé / livré / encaissé)",
        "description": "À appeler dès que la question porte sur l'argent gagné : « ce mois on a fait combien ? », « le chiffre d'affaires », « les recettes », « combien on a encaissé », « combien on a facturé », la comparaison entre deux mois. Rend TROIS chiffres DISTINCTS qu'il ne faut jamais additionner ni confondre : FACTURÉ (factures émises, brouillons exclus), LIVRÉ (commandes passées au statut Livrée) et ENCAISSÉ (recettes réellement passées par la caisse, d'après les rapports journaliers — c'est le chiffre que le Tableau de bord de l'application appelle « CA du mois »). Chaque chiffre porte sa définition : reprends-la si l'utilisateur demande pourquoi deux montants diffèrent. L'encaissé est en plus ventilé entre les deux activités, IMPRIMERIE OGOOUÉ et PAPETERIE OGOOUÉ.",
        "parameters": [
          {
            "name": "mois",
            "in": "query",
            "required": false,
            "description": "Mois voulu au format AAAA-MM, par exemple 2026-09. Omettre pour le mois en cours à Moanda.",
            "schema": { "type": "string", "pattern": "^\\d{4}-\\d{2}$" }
          }
        ],
        "responses": {
          "200": {
            "description": "Les trois chiffres du mois, datés.",
            "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } }
          }
        }
      }
    },
...
    "/api/chatgpt-collection": {
      "get": {
        "operationId": "lireUneCollection",
        "summary": "Lire les lignes d'une collection, avec pagination et filtre",
        "description": "À appeler pour lire le détail brut quand aucune synthèse ne répond : la fiche d'un client, les lignes d'une facture, les mouvements de stock, les pointages, le journal d'audit. Demande d'abord inventaireDesDonnees si tu ignores le nom exact de la collection. La lecture est paginée : commence par une petite limite, puis augmente `decalage` pour la suite plutôt que de tout rapatrier. Le filtre `champ` + `valeur` fait une égalité exacte sur un champ de la ligne (par exemple champ=statut et valeur=livree). Les empreintes de mots de passe ne sont jamais rendues, quelle que soit la collection.",
        "parameters": [
          {
            "name": "collection",
            "in": "query",
            "required": true,
            "description": "Nom exact de la collection, tel que rendu par inventaireDesDonnees (par exemple clients, factures, commandes, produits, employes, audit_logs).",
            "schema": { "type": "string" }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "description": "Nombre de lignes à rendre, de 1 à 200. Défaut 50.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }
          },
          {
            "name": "decalage",
            "in": "query",
            "required": false,
            "description": "Nombre de lignes à sauter, pour lire la page suivante. Défaut 0.",
            "schema": { "type": "integer", "minimum": 0, "default": 0 }
          },
          {
            "name": "champ",
            "in": "query",
            "required": false,
            "description": "Nom d'un champ de la ligne sur lequel filtrer (lettres, chiffres et tirets bas uniquement).",
            "schema": { "type": "string" }
          },
          {
            "name": "valeur",
            "in": "query",
            "required": false,
            "description": "Valeur exacte attendue pour le champ filtré. Sans effet si `champ` est absent.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Une page de lignes, le total et la pagination.",
            "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "jetonDuPont": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton du pont, à coller dans l'Action du GPT en type « API Key » / « Bearer ». Sa valeur est la variable d'environnement CHATGPT_BRIDGE_TOKEN définie sur Vercel. La changer dans Vercel coupe l'accès immédiatement."
      }
    }
  }
}